Move register button back to hero section
